completed/February_05_candidate_evaluation_system/requirements.txt
anthropic
python-dotenv
PyPDF2
